Created Casino
Casino will be just that, a variety of commands that mimics casino/gambling games

Roulette:
Roulette is the first implemented game
shows 9 roulette squares in a row.
- Randomness:
The board is offset by a random amount each time the command is run. The 'rolling' time is also for a random amount of time.